requirements.txt
# Core dependencies
pyautogen[contrib]>=0.2.0
requests>=2.31.0
python-dotenv>=1.0.0
flaml[automl]>=1.5.0